+++ packages/remax/src/adapters/toutiao/components/Canvas.ts
@@ -1,0 +1,9 @@
+import factory from './factory';
+
+export interface CanvasProps {
+  cavasId: string;
+}
+
+const Canvas = factory<CanvasProps>('canvas');
+
+export default Canvas;