+++ packages/remax/src/adapters/toutiao/components/Checkbox.ts
@@ -1,0 +1,12 @@
+import factory from './factory';
+
+export interface CheckboxProps {
+  value?: string;
+  disabled?: boolean;
+  checked?: boolean;
+  color?: string;
+}
+
+const Checkbox = factory<CheckboxProps>('checkbox');
+
+export default Checkbox;