+++ packages/remax/src/adapters/toutiao/components/Label.ts
@@ -1,0 +1,9 @@
+import factory from './factory';
+
+export interface LabelProps {
+  for?: string;
+}
+
+const Label = factory<LabelProps>('label');
+
+export default Label;