+++ packages/remax/src/adapters/toutiao/components/RadioGroup.ts
@@ -1,0 +1,9 @@
+import factory from './factory';
+
+export interface RadioGroupProps {
+  onChange?: (e: any) => void;
+}
+
+const RadioGroup = factory<RadioGroupProps>('radio-group');
+
+export default RadioGroup;