+++ packages/remax/typings/index.d.ts
@@ -1,1 +1,2 @@
 declare module 'scheduler';
+declare const tt;